refactor: extract main.rs god file into focused modules
Split the 1,837-line main.rs into a clean module structure:
- cli.rs: all clap struct definitions
- tracing_init.rs: tracing setup
- cmd/: one module per subcommand (hooks, init, status, explain,
  policy, launch, bug, commands, schema)

main.rs is now 72 lines of pure dispatch. No behavioral changes.
