+++ clash/src/cmd/mod.rs
@@ -1,0 +1,9 @@
+pub mod bug;
+pub mod commands;
+pub mod explain;
+pub mod hooks;
+pub mod init;
+pub mod launch;
+pub mod policy;
+pub mod schema;
+pub mod status;
